fix: id being incorrectly matched
adding id suffix to inputs to avoid duplicate id's lead to incorreclty matching and updating elements as we were comparing object ids which have no suffix to input ids which have suffix. Pass in the actual object id the input is referring to instead of the input id.

Also refactor custom hook to account for this and remove some duplicate code used in every form section